#272 move realisation into *tools module
