+++ proxy/core/acceptor/pool.py
@@ -27,6 +27,7 @@ logger = logging.getLogger(__name__)
 LOCK = multiprocessing.Lock()
 
 proxy_id_glob = multiprocessing.Value('i', 0)
+new_acc_id_glob = multiprocessing.Value('i', 0)
 
 
 class AcceptorPool:
